requirements-core.txt
# 核心預測管線 + 測試（CI / 最小安裝）
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
psutil>=5.9.0
